Initial version of the GanetiRapiClient class
Added the first version of the GanetiRapiClient class and a simple test
file to use with it. Most of the skeleton is done now. What we need to
do is code the different calls which can be found in the client.py from
the original project.
Useable methods currently written:
getVersion()
getFeatures()

TODO:
most of the other methods
cleaning to reflect what is really done in the original code.